refactor: move error summary component
