+++ src/app/carbon-estimator-form/carbon-estimator-form.component.ts
@@ -19,13 +19,13 @@ import {
   ErrorSummaryState,
 } from './carbon-estimator-form.constants';
 import { CloudFormSectionComponent } from '../features/cloud/components/cloud-form-section.component';
-import { ErrorSummaryComponent } from '../error-summary/error-summary.component';
 import { FormStateService } from '../services/form-state.service';
 import { SaasFormSectionComponent } from '../features/saas/components/saas-form-section.component';
 import { FormService } from '../services/form.service';
 import { OrganisationFormSectionComponent } from '../features/organisation/components/organisation-form-section.component';
 import { OnPremiseFormSectionComponent } from '../features/on-premise/components/on-premise-form-section.component';
 import { CustomerFormSectionComponent } from '../features/customers/components/customer-form-section.component';
+import { ErrorSummaryComponent } from '../components/error-summary/error-summary.component';
 @Component({
   selector: 'carbon-estimator-form',
   standalone: true,
